+++ prodotti_espresso/README.rst
@@ -1,12 +1,10 @@
 
-===================================
-|icon| Prodotti Espresso 10.0.1.0.5
-===================================
+==================
+|icon|  10.0.1.0.6
+==================
 
 
-**Prodotti espresso**
-
-.. |icon| image:: https://raw.githubusercontent.com/zeroincombenze/l10n-italy/10.0/prodotti_espresso/static/description/icon.png
+.. |icon| image:: https://raw.githubusercontent.com/zeroincombenze//10.0//static/description/icon.png
 
 |Maturity| |Build Status| |Codecov Status| |license gpl| |Try Me|
 
@@ -77,22 +75,23 @@ Installation / Installazione
     # Case 1: you have not installed zeroincombenze tools
     git clone https://github.com/zeroincombenze/tools.git
     cd $HOME/tools
-    ./install_tools.sh -p
+    ./install_tools.sh -pT
     source $HOME/devel/activate_tools
     # Case 2: you have already installed zeroincombenze tools
     cd $HOME/tools
-    ./install_tools.sh -U
+    ./install_tools.sh -UT
     source $HOME/devel/activate_tools
     # *** End of tools installation or upgrade ***
     # Odoo repository installation; OCB repository must be installed
-    odoo_install_repository l10n-italy -b 10.0 -O zero -o $HOME/10.0
-    vem create $HOME/10.0/venv_odoo -O 10.0 -a "*" -DI -o $HOME/10.0
+    deploy_odoo clone -r  -b 10.0 -G zero -p $HOME/10.0
+    # Upgrade virtual environment
+    vem amend $HOME/10.0/venv_odoo
 
 From UI: go to:
 
-* |menu| Setting > Activate Developer mode 
+* |menu| Setting > Activate Developer mode
 * |menu| Apps > Update Apps List
-* |menu| Setting > Apps |right_do| Select **prodotti_espresso** > Install
+* |menu| Setting > Apps |right_do| Select **** > Install
 
 
 |
@@ -108,16 +107,16 @@ Upgrade / Aggiornamento
     # Case 1: you have not installed zeroincombenze tools
     git clone https://github.com/zeroincombenze/tools.git
     cd $HOME/tools
-    ./install_tools.sh -p
+    ./install_tools.sh -pT
     source $HOME/devel/activate_tools
     # Case 2: you have already installed zeroincombenze tools
     cd $HOME/tools
-    ./install_tools.sh -U
+    ./install_tools.sh -UT
     source $HOME/devel/activate_tools
     # *** End of tools installation or upgrade ***
     # Odoo repository upgrade
-    odoo_install_repository l10n-italy -b 10.0 -o $HOME/10.0 -U
-    vem amend $HOME/10.0/venv_odoo -o $HOME/10.0
+    deploy_odoo update -r  -b 10.0 -G zero -p $HOME/10.0
+    vem amend $HOME/10.0/venv_odoo
     # Adjust following statements as per your system
     sudo systemctl restart odoo
 
@@ -125,7 +124,8 @@ From UI: go to:
 
 * |menu| Setting > Activate Developer mode
 * |menu| Apps > Update Apps List
-* |menu| Setting > Apps |right_do| Select **prodotti_espresso** > Update
+* |menu| Setting > Apps |right_do| Select **** > Update
+
 
 |
 
@@ -144,7 +144,7 @@ Get involved / Ci mettiamo in gioco
 
 Bug reports are welcome! You can use the issue tracker to report bugs,
 and/or submit pull requests on `GitHub Issues
-<https://github.com/zeroincombenze/l10n-italy/issues>`_.
+<https://github.com/zeroincombenze//issues>`_.
 
 In case of trouble, please check there if your issue has already been reported.
 
@@ -161,10 +161,17 @@ An Enhancement Proposal may be submitted if your idea gains ground.
 ChangeLog History / Cronologia modifiche
 ----------------------------------------
 
+12.0.1.0.6 (2023-05-30)
+~~~~~~~~~~~~~~~~~~~~~~~
+
+* [IMP] Flag espresso on sale order line form / Form righe ordini con flag prodotto espresso
+* [IMP] Flag espresso on sale order form / Form ordini con flag prodotto espresso
+* [IMP] Sale order BI with espresso filter / Filtro espresso in BI ordini
+
 12.0.1.0.5 (2023-02-08)
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-* [IMP] flag espresso on invoice line / righe fattura con flag prodotto espresso
+* [IMP] Flag espresso on invoice line / righe fattura con flag prodotto espresso
 
 12.0.1.0.4 (2023-01-20)
 ~~~~~~~~~~~~~~~~~~~~~~~
@@ -185,11 +192,6 @@ ChangeLog History / Cronologia modifiche
 ~~~~~~~~~~~~~~~~~~~~~~~
 
 * [IMP] Prodotto spedizione espresso
-
-12.0.1.0.0 (2022-05-25)
-~~~~~~~~~~~~~~~~~~~~~~~
-
-* [IMP] Impostazione modulo
 
 
 
@@ -213,8 +215,6 @@ Authors / Autori
 
 * `SHS-AV s.r.l. <https://www.zeroincombenze.it/>`__
 * `Didotech s.r.l. <https://www.didotech.com>`__
-Authors
--------
 
 
 Contributors / Collaboratori
@@ -223,9 +223,6 @@ Contributors / Collaboratori
 * Antonio Maria Vigliotti <antoniomaria.vigliotti@gmail.com>
 * Marco Tosato <marco.tosato@didotech.com>
 * Fabio Giovannelli <fabio.giovannelli@didotech.com>
-Contributors
-------------
-
 
 
 Maintainer / Manutenzione
@@ -255,15 +252,15 @@ La distribuzione `Zeroincombenze® <https://wiki.zeroincombenze.org/en/Odoo>`__ 
 
 |
 
-This module is part of l10n-italy project.
+This module is part of  project.
 
-Last Update / Ultimo aggiornamento: 2023-02-09
+Last Update / Ultimo aggiornamento: 
 
 .. |Maturity| image:: https://img.shields.io/badge/maturity-Beta-yellow.png
     :target: https://odoo-community.org/page/development-status
     :alt: 
-.. |Build Status| image:: https://travis-ci.org/zeroincombenze/l10n-italy.svg?branch=10.0
-    :target: https://travis-ci.com/zeroincombenze/l10n-italy
+.. |Build Status| image:: https://travis-ci.org/zeroincombenze/.svg?branch=10.0
+    :target: https://travis-ci.com/zeroincombenze/
     :alt: github.com
 .. |license gpl| image:: https://img.shields.io/badge/licence-LGPL--3-7379c3.svg
     :target: http://www.gnu.org/licenses/lgpl-3.0-standalone.html
@@ -271,11 +268,11 @@ Last Update / Ultimo aggiornamento: 2023-02-09
 .. |license opl| image:: https://img.shields.io/badge/licence-OPL-7379c3.svg
     :target: https://www.odoo.com/documentation/user/14.0/legal/licenses/licenses.html
     :alt: License: OPL
-.. |Coverage Status| image:: https://coveralls.io/repos/github/zeroincombenze/l10n-italy/badge.svg?branch=10.0
-    :target: https://coveralls.io/github/zeroincombenze/l10n-italy?branch=10.0
+.. |Coverage Status| image:: https://coveralls.io/repos/github/zeroincombenze//badge.svg?branch=10.0
+    :target: https://coveralls.io/github/zeroincombenze/?branch=10.0
     :alt: Coverage
-.. |Codecov Status| image:: https://codecov.io/gh/zeroincombenze/l10n-italy/branch/10.0/graph/badge.svg
-    :target: https://codecov.io/gh/zeroincombenze/l10n-italy/branch/10.0
+.. |Codecov Status| image:: https://codecov.io/gh/zeroincombenze//branch/10.0/graph/badge.svg
+    :target: https://codecov.io/gh/zeroincombenze//branch/10.0
     :alt: Codecov
 .. |Tech Doc| image:: https://www.zeroincombenze.it/wp-content/uploads/ci-ct/prd/button-docs-10.svg
     :target: https://wiki.zeroincombenze.org/en/Odoo/10.0/dev
@@ -286,8 +283,8 @@ Last Update / Ultimo aggiornamento: 2023-02-09
 .. |Try Me| image:: https://www.zeroincombenze.it/wp-content/uploads/ci-ct/prd/button-try-it-10.svg
     :target: https://erp10.zeroincombenze.it
     :alt: Try Me
-.. |OCA Codecov| image:: https://codecov.io/gh/OCA/l10n-italy/branch/10.0/graph/badge.svg
-    :target: https://codecov.io/gh/OCA/l10n-italy/branch/10.0
+.. |OCA Codecov| image:: https://codecov.io/gh/OCA//branch/10.0/graph/badge.svg
+    :target: https://codecov.io/gh/OCA//branch/10.0
     :alt: Codecov
 .. |Odoo Italia Associazione| image:: https://www.odoo-italia.org/images/Immagini/Odoo%20Italia%20-%20126x56.png
    :target: https://odoo-italia.org
@@ -317,4 +314,5 @@ Last Update / Ultimo aggiornamento: 2023-02-09
    :target: https://github.com/zeroincombenze/grymb/blob/master/certificates/ade/scope/fatturapa.md
 .. |chat_with_us| image:: https://www.shs-av.com/wp-content/chat_with_us.gif
    :target: https://t.me/Assitenza_clienti_powERP
+
