+++ prodotti_espresso/models/__init__.py
@@ -1,5 +1,6 @@
 # -*- coding: utf-8 -*-
 
 from . import product
+from . import account_invoice
 from . import sale_order
 from . import stock_picking_package_preparation
