+++ prodotti_espresso/report/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*-
 
 from . import account_invoice
+from . import sale_order
