+++ prodotti_espresso/report/account_invoice_report_view.xml
@@ -1,30 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <odoo>
 
-    <!--
-    <record id="view_account_invoice_report_espresso" model="ir.ui.view">
-        <field name="name">account.invoice.report.espresso</field>
-        <field name="model">account.invoice.report</field>
-        <field name="inherit_id" ref="account.view_account_invoice_report_pivot"/>
-        <field name="arch" type="xml">
-            <xpath expr="//field[@name='categ_id']" position="after">
-                <field name="espresso" type="row"/>
-            </xpath>
-        </field>
-    </record>
-
-    <record id="view_account_invoice_report_graph" model="ir.ui.view">
-        <field name="name">account.invoice.report.graph</field>
-        <field name="model">account.invoice.report</field>
-        <field name="inherit_id" ref="account.view_account_invoice_report_graph"/>
-        <field name="arch" type="xml">
-            <graph string="Invoices Analysis">
-                <field name="espresso"/>
-            </graph>
-        </field>
-    </record>
-    -->
-
     <record id="view_account_invoice_report_search" model="ir.ui.view">
         <field name="name">account.invoice.report.search</field>
         <field name="model">account.invoice.report</field>
