+++ prodotti_espresso/report/sale_order_report_view.xml
@@ -1,0 +1,15 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="view_sale_order_report_search" model="ir.ui.view">
+        <field name="name">sale.order.report.search</field>
+        <field name="model">sale.report</field>
+        <field name="inherit_id" ref="sale.view_order_product_search"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='categ_id']" position="after">
+                <filter name="Ordini espresso" domain="[('espresso', '=', True)]"/>
+            </xpath>
+        </field>
+    </record>
+
+</odoo>
