+++ prodotti_espresso/static/description/index.html
@@ -2,7 +2,7 @@
 <section class="oe_container">
     <div class="oe_row oe_spaced">
         <div class="oe_span12">
-            <h2 class="oe_slogan" style="color:#162832;">Prodotti espresso / Questo modulo permette di creare dei ddt dagli ordini di vendita con prodotti "espresso" raggruppati per indirizzo di spedizione.</h2>
+            <h2 class="oe_slogan" style="color:#162832;"> / Questo modulo permette di creare dei ddt dagli ordini di vendita con prodotti "espresso" raggruppati per indirizzo di spedizione.</h2>
         </div>
         <div class="oe_span6">
             <p align="center">
@@ -20,10 +20,18 @@
     <div class="oe_row oe_spaced">
         <h3 class="oe_slogan" style="color:#162832;">History / Cronologia</h3>
         <div class="oe_span12">
-<h3>12.0.1.0.5 (2023-02-08)</h3>
+<h3>12.0.1.0.6 (2023-05-30)</h3>
 <p align="justify">
 <ul>
-<li>[IMP] flag espresso on invoice line / righe fattura con flag prodotto espresso</li>
+<li>[IMP] Flag espresso on sale order line form / Form righe ordini con flag prodotto espresso</li>
+<li>[IMP] Flag espresso on sale order form / Form ordini con flag prodotto espresso</li>
+<li>[IMP] Sale order BI with espresso filter / Filtro espresso in BI ordini</li>
+</ul>
+</p><p align="justify">
+<h3>12.0.1.0.5 (2023-02-08)</h3>
+</p><p align="justify">
+<ul>
+<li>[IMP] Flag espresso on invoice line / righe fattura con flag prodotto espresso</li>
 </ul>
 </p><p align="justify">
 <h3>12.0.1.0.4 (2023-01-20)</h3>
@@ -49,12 +57,6 @@
 <ul>
 <li>[IMP] Prodotto spedizione espresso</li>
 </ul>
-</p><p align="justify">
-<h3>12.0.1.0.0 (2022-05-25)</h3>
-</p><p align="justify">
-<ul>
-<li>[IMP] Impostazione modulo</li>
-</ul>
 </p>
         </div>
     </div>
@@ -69,8 +71,7 @@
             <p class="oe_mt16"><ul>
 <li><a href="https://www.zeroincombenze.it/">SHS-AV s.r.l.</a></li>
 <li><a href="https://www.didotech.com/">Didotech s.r.l.</a></li>
-</ul>
-<h2>Prodotti espresso / Questo modulo permette di creare dei ddt dagli ordini di vendita con prodotti "espresso" raggruppati per indirizzo di spedizione.</h2></p>
+</ul></p>
         </div>
         <div class="oe_span6" align="left">
             <p class="oe_mt16">Contributors / Collaboratori:</p>
@@ -78,8 +79,7 @@
 <li>Antonio Maria Vigliotti &lt;antoniomaria.vigliotti@gmail.com&gt;</li>
 <li>Marco Tosato &lt;marco.tosato@didotech.com&gt;</li>
 <li>Fabio Giovannelli &lt;fabio.giovannelli@didotech.com&gt;</li>
-</ul>
-<h2>Prodotti espresso / Questo modulo permette di creare dei ddt dagli ordini di vendita con prodotti "espresso" raggruppati per indirizzo di spedizione.</h2></p>
+</ul></p>
         </div>
     </div>
     <div class="oe_row oe_spaced">
@@ -88,6 +88,6 @@
         </div>
     </div>
     <div class="oe_row oe_spaced">
-        <p>This module is part of l10n-italy project.</p>
+        <p>This module is part of  project.</p>
     </div>
 </section>
