+++ prodotti_espresso/tests/logs/prodotti_espresso_10.sh
@@ -1,1 +1,1 @@
-cd /home/odoo/10.0; /home/odoo/10.0/odoo-bin --stop-after-init -d test_odoo_10 --config=/home/odoo/10.0/l10n-italy/prodotti_espresso/tests/logs/zero10.l10n-italy.prodotti_espresso.conf --log-level=error -i account,account_invoice_pricelist,analytic,barcodes,base,base_setup,bus,decimal_precision,delivery,l10n_it_ade,l10n_it_ddt,mail,mrp,procurement,product,report,resource,sale,sale_stock,sales_team,stock,stock_account,stock_picking_package_preparation,stock_picking_package_preparation_line,web,web_kanban,web_planner,web_tour
+cd /home/odoo/10.0; /home/odoo/10.0/odoo-bin --stop-after-init -d test_odoo_10 --config=/home/odoo/10.0/l10n-italy/prodotti_espresso/tests/logs/zero10.l10n-italy.prodotti_espresso.conf --log-level=error -i account,account_invoice_partner_carrier,account_invoice_pricelist,analytic,barcodes,base,base_setup,bus,decimal_precision,delivery,l10n_it_ade,l10n_it_ddt,mail,mrp,procurement,product,report,resource,sale,sale_stock,sales_team,stock,stock_account,stock_picking_package_preparation,stock_picking_package_preparation_line,web,web_kanban,web_planner,web_tour
