+++ prodotti_espresso/tests/logs/requirements.txt
@@ -1,17 +1,24 @@
+AccessControl==4.3
 acme-tiny==5.0.1
+Acquisition==4.13
 asn1crypto==1.5.1
-astroid==1.6.6
-Babel==2.9.1
+astroid==1.6.5
+atomicwrites==1.4.1
+attrs==21.4.0
+AuthEncoding==4.3
+Babel==2.3.4
 backports.functools-lru-cache==1.6.4
 bcrypt==3.1.7
 beautifulsoup4==4.9.3
+BTrees==4.11.3
 cachetools==3.1.1
 certifi==2021.10.8
 cffi==1.15.1
+Chameleon==3.10.2
 chardet==4.0.0
 checksumdir==1.1.7
 click==7.1.2
-clodoo==2.0.2
+clodoo==2.0.4
 codecov==2.1.12
 codicefiscale==0.9
 configparser==4.0.2
@@ -20,76 +27,91 @@ coverage==5.5
 coveralls==1.11.1
 cryptography==3.3.2
 cssselect==1.1.0
+DateTime==5.0
 decorator==4.0.10
 dnspython==1.16.0
 docopt==0.6.2
-docutils==0.14
+DocumentTemplate==3.4
+docutils==0.16
 ebaysdk==2.1.4
+email-validator==1.2.1
 enum34==1.1.10
 et-xmlfile==1.0.1
+ExtensionClass==4.9
 fdb==2.0.2
 feedparser==5.2.1
-flake8==3.9.2
+flake8==3.4.1
 funcsigs==1.0.2
 functools32==3.2.3.post2
-future==0.18.2
+future==0.18.3
 futures==3.4.0
 gdata==2.0.18
 Genshi==0.7.7
-gevent==1.1.2
+gevent==22.10.2
 gitdb2==2.0.6
 GitPython==2.1.15
-greenlet==0.4.10
+greenlet==2.0.2
 html2text==2019.8.11
 idna==2.10
 importlib-metadata==2.1.3
 ipaddress==1.0.23
 IPy==1.1
-isort==4.3.21
+isort==4.3.4
 jcconv==0.2.3
 jdcal==1.4.1
 Jinja2==2.10.1
 jsonlib==1.6.1
 lazy-object-proxy==1.6.0
-lxml==4.9.1
-Mako==1.0.4
-MarkupSafe==0.23
+lxml==4.9.2
+Mako==1.1.6
+MarkupSafe==1.1.1
 mccabe==0.6.1
 mock==2.0.0
+more-itertools==5.0.0
+MultiMapping==4.1
+multipart==0.1.1
 num2words==0.5.12
 numpy==1.16.6
 odoo==10.0
-odoo-score==2.0.3
+odoo-score==2.0.5
 OdooRPC==0.9.0
 OERPLib==0.8.4
 ofxparse==0.16
 openpyxl==2.6.4
-openupgradelib==3.3.5.dev24+gf0b809f
+openupgradelib==3.4.2.dev4+gf45dcfe
 os0==2.0.1
+packaging==20.9
 paramiko==2.12.0
 passlib==1.6.5
+PasteDeploy==2.1.1
 pathlib2==2.3.7.post1
-pbr==5.11.0
+pbr==5.11.1
+pep8==1.7.1
+Persistence==3.6
+persistent==4.9.3
 Pillow==3.4.1
 pip==20.3.4
 pkcs7==0.1.2
-pkg-resources==0.0.0
-polib==1.1.1
+pluggy==0.13.1
+polib==1.1.0
 psutil==4.3.1
 psycogreen==1.0
 psycopg2-binary==2.8.6
+py==1.11.0
 py3o.formats==0.3
 py3o.template==0.10.0
 pyasn1==0.4.8
-pycodestyle==2.7.0
+pyasn1-modules==0.2.8
+pycodestyle==2.3.1
 pycparser==2.21
 pycrypto==2.6.1
 pydot==1.2.3
-pyflakes==2.3.1
+pyflakes==1.5.0
 Pygments==2.2.0
 pyjon.utils==0.7
 pylint==1.9.5
-pylint-odoo==5.0.0
+pylint-mccabe==0.1.3
+pylint-odoo==3.5.0
 pylint-plugin-utils==0.6
 PyNaCl==1.4.0
 pyOpenSSL==21.0.0
@@ -99,52 +121,113 @@ pyPdf==1.13
 PyPDF2==1.28.4
 pyserial==3.5
 pysftp==0.2.9
+pytest==4.6.11
 Python-Chart==1.39
 python-dateutil==2.5.3
+python-gettext==4.1
 python-ldap==2.4.27
 python-magic==0.4.27
 python-openid==2.2.5
-python-plus==2.0.2.2
-python-stdnum==1.1
-pytz==2016.7
-pyusb==1.1.0
+python-plus==2.0.5
+python-stdnum==1.18
+pytz==2022.7.1
+pyusb==1.0.0
+PyWebDAV==0.9.8
 PyXB==1.2.5
 PyYAML==3.12
 qrcode==5.3
+QUnitSuite==0.4
 raven==6.10.0
 reportlab==3.3.0
 requests==2.27.1
-restructuredtext-lint==1.3.2
+RestrictedPython==5.2
+restructuredtext-lint==1.3.1
 rfc3986==1.5.0
+roman==3.3
 scandir==1.10.0
 setuptools==44.1.1
-simplejson==3.5.3
+simplejson==3.18.3
 singledispatch==3.7.0
-six==1.10.0
+six==1.16.0
 smmap==3.0.5
 smmap2==3.0.1
 soupsieve==1.9.6
-SQLAlchemy==1.4.43
-suds==0.4
+SQLAlchemy==1.4.46
 suds-jurko==0.6
+transaction==3.0.1
 typing==3.10.0.0
 unicodecsv==0.14.1
 Unidecode==1.2.0
 unittest2==0.5.1
-urllib3==1.26.12
+urllib3==1.26.14
 urllib3-secure-extra==0.1.0
 vatnumber==1.2
 viivakoodi==0.8.0
 vobject==0.9.3
+waitress==1.4.4
+wcwidth==0.2.6
+WebOb==1.8.7
+websocket-client==0.59.0
+WebTest==2.0.35
 Werkzeug==0.11.11
-wheel==0.34.2
+wheel==0.37.1
 whichcraft==0.6.1
-wrapt==1.14.1
+wrapt==1.15.0
+WSGIProxy2==0.5.1
 xlrd==1.0.0
 XlsxWriter==0.9.3
 xlwt==1.1.2
-z0bug-odoo==2.0.1.1
-z0lib==2.0.1
-zerobug==2.0.2
+z0bug-odoo==2.0.6
+z0lib==2.0.3
+z3c.pt==3.3.1
+zc.lockfile==2.0
+ZConfig==3.6.1
+zerobug==2.0.4
+zExceptions==4.3
 zipp==1.2.0
-zpl2==1.0
+ZODB==5.8.0
+zodbpickle==2.6
+Zope==4.8.7
+zope.annotation==4.8
+zope.browser==2.4
+zope.browsermenu==4.4
+zope.browserpage==4.4.0
+zope.browserresource==4.4
+zope.cachedescriptors==4.4
+zope.component==5.1.0
+zope.configuration==4.4.1
+zope.container==4.10
+zope.contentprovider==4.2.1
+zope.contenttype==4.6
+zope.datetime==4.3.0
+zope.deferredimport==4.4
+zope.deprecation==4.4.0
+zope.dottedname==5.0
+zope.event==4.6
+zope.exceptions==4.6
+zope.filerepresentation==5.0.0
+zope.globalrequest==1.6
+zope.hookable==5.4
+zope.i18n==4.9.0
+zope.i18nmessageid==5.1.1
+zope.interface==5.5.2
+zope.lifecycleevent==4.4
+zope.location==4.3
+zope.pagetemplate==4.6.0
+zope.processlifetime==2.4
+zope.proxy==4.6.1
+zope.ptresource==4.3.0
+zope.publisher==6.1.0
+zope.schema==6.2.1
+zope.security==5.8
+zope.sequencesort==4.2
+zope.site==4.6.1
+zope.size==4.4
+zope.structuredtext==4.4
+zope.tal==4.5
+zope.tales==5.2
+zope.testbrowser==5.6.1
+zope.testing==5.0.1
+zope.traversing==4.4.1
+zope.viewlet==4.3
+zpl2==1.2.1
