+++ prodotti_espresso/tests/logs/show-log.sh
@@ -1,1 +1,1 @@
-less -R $(readlink -f $(dirname $0))/prodotti_espresso_20230209.txt
+less -R $(readlink -f $(dirname $0))/prodotti_espresso_20230531.txt
