+++ prodotti_espresso/tests/logs/zero10.l10n-italy.prodotti_espresso.conf
@@ -1,5 +1,5 @@
 [options]
-addons_path = /home/odoo/10.0/odoo/addons,/home/odoo/10.0/addons,/home/odoo/10.0/account-analytic,/home/odoo/10.0/account-budgeting,/home/odoo/10.0/account-closing,/home/odoo/10.0/account-consolidation,/home/odoo/10.0/account-financial-reporting,/home/odoo/10.0/account-financial-tools,/home/odoo/10.0/account-fiscal-rule,/home/odoo/10.0/account-invoice-reporting,/home/odoo/10.0/account-invoicing,/home/odoo/10.0/account-payment,/home/odoo/10.0/account-reconcile,/home/odoo/10.0/apps-store,/home/odoo/10.0/bank-payment,/home/odoo/10.0/bank-statement-import,/home/odoo/10.0/business-requirement,/home/odoo/10.0/calendar,/home/odoo/10.0/commission,/home/odoo/10.0/community-data-files,/home/odoo/10.0/connector,/home/odoo/10.0/contract,/home/odoo/10.0/credit-control,/home/odoo/10.0/crm,/home/odoo/10.0/currency,/home/odoo/10.0/data-protection,/home/odoo/10.0/delivery-carrier,/home/odoo/10.0/donation,/home/odoo/10.0/e-commerce,/home/odoo/10.0/edi,/home/odoo/10.0/event,/home/odoo/10.0/geospatial,/home/odoo/10.0/helpdesk,/home/odoo/10.0/hr,/home/odoo/10.0/infrastructure,/home/odoo/10.0/intrastat-extrastat,/home/odoo/10.0/knowledge,/home/odoo/10.0/l10n-italy,/home/odoo/10.0/l10n-italy-supplemental,/home/odoo/10.0/maintenance,/home/odoo/10.0/management-system,/home/odoo/10.0/manufacture,/home/odoo/10.0/manufacture-reporting,/home/odoo/10.0/margin-analysis,/home/odoo/10.0/mgmtsystem,/home/odoo/10.0/multi-company,/home/odoo/10.0/operating-unit,/home/odoo/10.0/partner-contact,/home/odoo/10.0/pos,/home/odoo/10.0/product-attribute,/home/odoo/10.0/product-kitting,/home/odoo/10.0/product-variant,/home/odoo/10.0/profiles,/home/odoo/10.0/project,/home/odoo/10.0/project-agile,/home/odoo/10.0/project-reporting,/home/odoo/10.0/purchase-reporting,/home/odoo/10.0/purchase-workflow,/home/odoo/10.0/queue,/home/odoo/10.0/report-print-send,/home/odoo/10.0/reporting-engine,/home/odoo/10.0/rma,/home/odoo/10.0/sale-financial,/home/odoo/10.0/sale-reporting,/home/odoo/10.0/sale-workflow,/home/odoo/10.0/search-engine,/home/odoo/10.0/server-auth,/home/odoo/10.0/server-tools,/home/odoo/10.0/server-ux,/home/odoo/10.0/social,/home/odoo/10.0/stock-logistics-barcode,/home/odoo/10.0/stock-logistics-reporting,/home/odoo/10.0/stock-logistics-tracking,/home/odoo/10.0/stock-logistics-transport,/home/odoo/10.0/stock-logistics-warehouse,/home/odoo/10.0/stock-logistics-workflow,/home/odoo/10.0/storage,/home/odoo/10.0/timesheet,/home/odoo/10.0/uncovered,/home/odoo/10.0/web,/home/odoo/10.0/webhook,/home/odoo/10.0/webkit-tools,/home/odoo/10.0/website,/home/odoo/10.0/website-cms,/home/odoo/10.0/wms,/home/odoo/10.0/zerobug-test
+addons_path = /home/odoo/10.0/odoo/addons,/home/odoo/10.0/addons,/home/odoo/10.0/account-analytic,/home/odoo/10.0/account-budgeting,/home/odoo/10.0/account-closing,/home/odoo/10.0/account-consolidation,/home/odoo/10.0/account-financial-reporting,/home/odoo/10.0/account-financial-tools,/home/odoo/10.0/account-fiscal-rule,/home/odoo/10.0/account-invoice-reporting,/home/odoo/10.0/account-invoicing,/home/odoo/10.0/account-payment,/home/odoo/10.0/account-reconcile,/home/odoo/10.0/apps-store,/home/odoo/10.0/bank-payment,/home/odoo/10.0/bank-statement-import,/home/odoo/10.0/business-requirement,/home/odoo/10.0/calendar,/home/odoo/10.0/commission,/home/odoo/10.0/community-data-files,/home/odoo/10.0/connector,/home/odoo/10.0/contract,/home/odoo/10.0/credit-control,/home/odoo/10.0/crm,/home/odoo/10.0/currency,/home/odoo/10.0/data-protection,/home/odoo/10.0/delivery-carrier,/home/odoo/10.0/donation,/home/odoo/10.0/e-commerce,/home/odoo/10.0/edi,/home/odoo/10.0/event,/home/odoo/10.0/geospatial,/home/odoo/10.0/helpdesk,/home/odoo/10.0/hr,/home/odoo/10.0/infrastructure,/home/odoo/10.0/intrastat-extrastat,/home/odoo/10.0/knowledge,/home/odoo/10.0/l10n-italy,/home/odoo/10.0/l10n-italy-supplemental,/home/odoo/10.0/maintenance,/home/odoo/10.0/management-system,/home/odoo/10.0/manufacture,/home/odoo/10.0/manufacture-reporting,/home/odoo/10.0/margin-analysis,/home/odoo/10.0/mgmtsystem,/home/odoo/10.0/multi-company,/home/odoo/10.0/operating-unit,/home/odoo/10.0/partner-contact,/home/odoo/10.0/pos,/home/odoo/10.0/product-attribute,/home/odoo/10.0/product-kitting,/home/odoo/10.0/product-variant,/home/odoo/10.0/profiles,/home/odoo/10.0/project,/home/odoo/10.0/project-agile,/home/odoo/10.0/project-reporting,/home/odoo/10.0/purchase-reporting,/home/odoo/10.0/purchase-workflow,/home/odoo/10.0/queue,/home/odoo/10.0/report-print-send,/home/odoo/10.0/reporting-engine,/home/odoo/10.0/rma,/home/odoo/10.0/sale-financial,/home/odoo/10.0/sale-reporting,/home/odoo/10.0/sale-workflow,/home/odoo/10.0/search-engine,/home/odoo/10.0/server-auth,/home/odoo/10.0/server-tools,/home/odoo/10.0/server-ux,/home/odoo/10.0/social,/home/odoo/10.0/stock-logistics-barcode,/home/odoo/10.0/stock-logistics-reporting,/home/odoo/10.0/stock-logistics-tracking,/home/odoo/10.0/stock-logistics-transport,/home/odoo/10.0/stock-logistics-warehouse,/home/odoo/10.0/stock-logistics-workflow,/home/odoo/10.0/storage,/home/odoo/10.0/timesheet,/home/odoo/10.0/uncovered,/home/odoo/10.0/web,/home/odoo/10.0/webhook,/home/odoo/10.0/webkit-tools,/home/odoo/10.0/website,/home/odoo/10.0/website-cms,/home/odoo/10.0/wms,/home/odoo/10.0/zerobug-test,/home/odoo/10.0/social
 admin_passwd = shs20!Z1
 csv_internal_sep = ,
 data_dir = /home/odoo/.local/share/Odoo10
@@ -55,6 +55,6 @@ without_demo = False
 workers = 0
 xmlrpc = True
 xmlrpc_interface = 
-xmlrpc_port = 30321
+xmlrpc_port = 64076
