+++ prodotti_espresso/views/account_invoice_view.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+        <record id="invoice_form_view_conai" model="ir.ui.view">
+        <field name="name">account.invoice.form.espresso</field>
+        <field name="model">account.invoice</field>
+        <field name="inherit_id" ref="account.invoice_form"/>
+        <field name="arch" type="xml">
+
+            <field name="date_due" position="after">
+                <field name="espresso"/>
+            </field>
+
+        </field>
+    </record>
+
+    <record id="invoice_line_form_view_espresso" model="ir.ui.view">
+        <field name="name">account.invoice.line_form.espresso</field>
+        <field name="model">account.invoice.line</field>
+        <field name="inherit_id" ref="account.view_invoice_line_form"/>
+        <field name="arch" type="xml">
+
+            <field name="invoice_line_tax_ids" position="after">
+                <field name="espresso"/>
+            </field>
+
+        </field>
+    </record>
+
+</odoo>
