+++ prodotti_espresso/views/sale_order_view.xml
@@ -1,0 +1,40 @@
+<?xml version="1.0" encoding="utf-8"?>
+<odoo>
+
+    <record id="sale_order_form_view_espresso" model="ir.ui.view">
+        <field name="name">sale.order.form.espresso</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_order_form"/>
+        <field name="arch" type="xml">
+            <xpath expr="//field[@name='project_id']" position="after">
+                <field name="espresso"/>
+            </xpath>
+            <xpath
+                expr="//field[@name='order_line']//field[@name='tax_id']"
+                position="after"
+            >
+                <field name="espresso"/>
+            </xpath>
+        </field>
+    </record>
+
+        <!--
+    <record id="sale_order_search" model="ir.ui.view">
+        <field name="name">sale.order.search</field>
+        <field name="model">sale.order</field>
+        <field name="inherit_id" ref="sale.view_sales_order_filter"/>
+        <field name="arch" type="xml">
+            <xpath expr="//filter[@name='my_sale_orders_filter']" position="after">
+                <filter
+                    string="Documenti con prodotti espresso"
+                    domain="[('espresso', '!=', False)]"
+                />
+            </xpath>
+            <search>
+                <field name="espresso"/>
+            </search>
+        </field>
+    </record>
+        -->
+
+</odoo>
